requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
yt-dlp>=2024.1.1
sqlalchemy>=2.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
pywebview>=4.4.0
pyinstaller>=6.0.0
requests-oauthlib>=1.3.1
tls-client>=0.2.2